restfulapi/f1ed-t2_whosbug-restfulapi-master/requirements.txt
Django==2.2      
django-filter==2.3.0    
djangorestframework==3.11.0 
psycopg2
psycopg2-binary==2.8.5
